Remove MPEG Working Draft on Carriage of Web Resources in ISO BMFF
This is a draft document in development at MPEG. It describes a
method for enriching video content with synchronized interactive
web content. As we want the present document to focus on
requirements specifically for DataCue and timed event synchronization,
and the MPEG dfaft goes much further than this, it seems appropriate
to remove the reference at the present time. Further discussion on
this MPEG draft is welcome.