Rewrite gap analysis and recommendations for sync
The rewrite re-shuffles and completes the text that talked about synchronization
in the gap analysis section to:
- bring together mechanisms that web apps can use as subsections within the
same "synchronized rendering of web resources", without distinguishing between
text track cues and other non media web content
- clarify the synchronization challenge, notably due to rendering of media
content following a different path and clock
- list specific needs in generic terms, not linked to the notion of cues
- distinguish between progress along the media timeline and detection of a
frame boundary (needed in some cases but not all)
- add a recommendation for a mechanism that would allow to detect frame
boundaries when that is needed.

The update also adds more terms to the terminology section.